requirements.txt
django
pillow
djangorestframework
drf-yasg
# django-filter
django-debug-toolbar
